renamed and template class written in one file
